requirements.txt
requests
websocket-client
python-dotenv
cloudscraper
pyttsx3
PySide6